WEAT-5 Improve tests readability
