+++ app/src/test/java/com/app/weather/presentation/weather/FakeWeatherRepository.kt
@@ -1,14 +1,10 @@
-package com.app.weather.presentation
+package com.app.weather.presentation.weather
 
 import com.app.core.data.repository.weather.WeatherRepository
-import com.app.core.domain.*
+import com.app.core.domain.ResultWrapper
 import com.app.core.domain.weather.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-
-/**
- * Created by alirezarafeezadeh on 7/2/21.
- */
 
 class FakeWeatherRepository : WeatherRepository {
     override suspend fun forecast(latLong: String): Flow<ResultWrapper<ForecastResponse>> = flow {
@@ -22,8 +18,6 @@ class FakeWeatherRepository : WeatherRepository {
 
         val location = Location("", 1.0, "", 1, 1.0, "",
         "", "")
-
-//        ForecastResponse(current, Forecast(emptyList()), location)
         emit(ResultWrapper.Success(ForecastResponse(current, Forecast(emptyList()), location)))
     }
 }
