+++ core/src/test/java/com/app/core/repository/weather/WeatherRepositoryImplTest.kt
@@ -3,8 +3,8 @@ package com.app.core.repository.weather
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.app.core.MainCoroutineRule
 import com.app.core.data.repository.weather.WeatherRepositoryImpl
-import com.app.core.domain.*
-import com.app.core.domain.weather.*
+import com.app.core.domain.ResultWrapper
+import com.app.core.repository.util.ForecastMockData.forecastFakeResponse
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.collect
@@ -12,7 +12,6 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-
 
 @ExperimentalCoroutinesApi
 class WeatherRepositoryImplTest {
@@ -34,25 +33,8 @@ class WeatherRepositoryImplTest {
     fun forecast() = runBlocking {
         val response = weatherRepository.forecast("")
         val expectedResp = forecastFakeResponse()
-
         response.collect {
-//            ResultWrapper.Success(expectedResp)
             assertThat(it).isEqualTo(ResultWrapper.Success(expectedResp))
         }
     }
-
-    private fun forecastFakeResponse(): ForecastResponse {
-        val current = Current(
-            1, Condition(0,"",""), 1.0, 1.0, 1.0,
-            1.0, 1, 1, "", 1, 1.0,
-            1.0, 1.0, 1.0, 1.0, 1.0, 1.0,
-            1.0, 1.0, 1, "",1.0,1.0
-        )
-
-        val location = Location("",1.0,"",1,1.0,"",
-            "","")
-
-        return ForecastResponse(current, Forecast(emptyList()), location)
-    }
-
 }
